Move beam specific folders with scripts into separate `scripts` folder
